# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Recall API
  version: "1"
  description: |
    HTTP REST + SSE surface for the Recall Overwatch metrics app.

    Available in two flavors:
      * **Server binary** (`Recall-server`) — exposes this API by default
      * **Wails desktop binary with `--server`** — same API, same routes

    Conventions:
      * Every JSON endpoint is mounted under the `/api/v1/` prefix —
        the version is bumped on breaking changes.
      * Resources are nouns; sub-resources hang off the parent
        (`/matches/{match_key}/visibility`, `/settings/tesseract`).
      * Methods reflect intent: `GET` to read, `PUT` to replace/upsert,
        `DELETE` to clear or wipe, `POST` to start a server-side action
        that doesn't map to a single resource (the parse run).
      * Writes return `204 No Content` when there is no useful body
        to echo back; `202 Accepted` for actions whose effect is
        asynchronous (parse).
      * Errors come back as `application/problem+json` (RFC 9457) with
        the appropriate 4xx/5xx status. `type` is a stable URI under
        `https://github.com/sound-barrier/recall/problems/{slug}` that
        clients can switch on; body-validation problems carry a
        per-field `errors` array, and the data-update 422 carries a
        `failed_assets` list (RFC 9457 §3.2 extension members). The one
        exception is `405 Method Not Allowed`, emitted plain by the
        router.
...

...
    `RECALL_SERVER_ADDR`.
  license:
...
servers:
...
paths:
  /api/v1/matches:
    get:
      operationId: GetMatchResults
      tags: [Matches]
      summary: List parsed match records (optionally paginated)
      description: |
        Returns match records, assembled by JOINing the
        per-screenshot-type rows (summary / teams / personal /
        rank / unknown) that share each `match_key`. Read-time
        derivations (`inferSoleHeroPercent`, `inferResultFromRank`)
        plus role-from-hero and type-from-map lookups are applied
        on the fly — the underlying rows are never mutated.

        **Pagination.** Optional `limit` (1–1000) caps the number
        of records returned. Optional `cursor` is the previous
        page's last `match_key`; the next page starts strictly
        after it. Omitting both returns the full corpus
        (back-compat). The cursor is the match_key itself, not an
        opaque encoded value, so curl users can page without a
        lookup table.
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
          description: Max records per page. Clamped to [1, 1000].
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: |
            The previous page's last `match_key`. Returned records
            start strictly after it. Empty / absent = first page.
      responses:
        "200":
          description: Array of match records.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchRecord"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: ClearMatches
      tags: [Matches]
      summary: Wipe all parsed match records
      description: |
        Deletes every row from the per-screenshot tables (and their
        child rows via `ON DELETE CASCADE`) plus the auxiliary tables
        (reviews, queue, play-mode, annotations, hidden flags,
        ambiguous candidates, and the suppress-list). Settings and
        the screenshots folder are untouched; you can re-parse to
        rebuild.

        The optional `keep_ignored=true` query param preserves the
        Unknown-tab "Delete forever" suppress list across the wipe
        (Settings → Advanced surfaces this as a "Keep suppress-list"
        checkbox). Default-off matches the historical factory-reset
        semantic.

        `keep_ignored` covers only that suppress list — the separate
        "All Heroes" recognition skip-list is always cleared, even with
        `keep_ignored=true`. This is benign: those screenshots are
        re-recognized and re-skipped on the next parse, so nothing is
        lost.
      parameters:
        - name: keep_ignored
          in: query
          required: false
          schema:
            type: boolean
          description: |
            Preserve the `ignored_screenshots` suppress list across
            the wipe. Defaults to false (the historical factory-reset
            semantic). Strict — anything other than `true` / `false`
            / absent returns 400.
      responses:
        "204":
          description: All matches cleared.
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: CreateManualMatch
      tags: [Matches]
      summary: Hand-enter a match (no OCR)
      description: |
        Creates a manual match for users without Tesseract. The server
        derives the `match_key` from `played_at` (default now), rejects a
        collision with any existing match (409 — pick a different minute),
        writes the user-data override row plus the queue / play-mode aux
        rows, and returns the aggregated MatchRecord (`source: manual`).
        The right-side detail-panel choosers (review, replay code, …) then
        work unchanged, keyed by the new `match_key`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualMatchInput"
      responses:
        "201":
          description: The created manual match.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchRecord"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: A match already exists at that time; pick a different minute.
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/matches/{match_key}:
    parameters:
      - $ref: "#/components/parameters/MatchKey"
    get:
      operationId: GetMatchByKey
      tags: [Matches]
      summary: Read a single match by key
      description: |
        Returns the aggregated `MatchRecord` for one match. Identical
        wire shape to a single entry from `GetMatchResults`, with the
        same read-time inference + child-table folding semantics —
        consumers that need just one match can avoid pulling the
        whole corpus.

        404 on unknown keys (the `ErrMatchNotFound` sentinel routes
        through `errors.Is` to keep honest "not in corpus" out of
        the 500 catch-all).
      responses:
        "200":
          description: Match found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchRecord"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: Match not found.
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: HardDeleteMatch
      tags: [Matches]
      summary: Hard-delete a single match
      description: |
        Removes every trace of the match from the database: rows
        across all five parent screenshot tables (children CASCADE),
        the `match_annotations` row, and the `hidden_matches` flag.
        Source PNG files on disk are NOT touched, but `LoadAllFilenames`
        will then re-classify them as new and the next parse will
        re-ingest. Surfaced by the Hidden drawer's **Delete forever**
        affordance — the user has already moved the match to the
        archive and explicitly confirmed.

        Idempotent: unknown keys return 204.
      responses:
        "204":
          description: Match removed (or already absent).
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/v1/matches/{match_key}/resolution:
    parameters:
      - $ref: "#/components/parameters/MatchKey"
    put:
      operationId: ResolveAmbiguousMatch
      tags: [Matches]
      summary: Resolve an ambiguous-attribution screenshot
      description: |
        When the resolver can't pin a screenshot to a single match
        (identical EAD signature within the 5-30 min ambiguous window,
        or multiple candidates within 30 min), the parent row's
        `match_key` is set to `ambiguous-<filename>` and the candidate
        list is recorded in `ambiguous_candidates`. Other screenshots
        captured within `mergeWindow` of that one adopt the same
        sentinel via the timestamp-window pass — so several rows can
        share one ambiguous `match_key`.

        This endpoint rewrites every parent row carrying the given
        ambiguous key to `resolved_to`. The path `match_key` must
        start with `ambiguous-` (anything else returns 400).
        `resolved_to` must
        either be one of the recorded candidates OR a freshly-minted
        `match-<...>` key (the "Treat as new match" escape hatch in
        the Unknown tab).

        On success the ambiguous row + its candidates cascade-delete
        and a `match-updated` SSE event fires for the resolved key.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resolved_to]
              properties:
                resolved_to:
                  type: string
                  description: |
                    The match the user picked. Must be one of the
                    recorded candidates or a `match-<...>` key.
                  example: "match-2026-05-10T21-29-28"
      responses:
        "204":
          description: Resolution applied.
        "400":
          description: |
            Invalid input — `match_key` doesn't start with
            `ambiguous-`, `resolved_to` is not one of the recorded
            candidates and isn't a freshly-minted `match-<...>` key,
            or the request body fails JSON validation.
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/matches/{match_key}/annotation:
    parameters:
      - $ref: "#/components/parameters/MatchKey"
    put:
      operationId: SetMatchAnnotation
      tags: [Matches]
      summary: Upsert a per-match user annotation
      description: |
        Lets the user tag a match with up to four pieces of metadata:

          - **leaver** — `self` (user left, data incomplete) /
            `team` (ally left, loss excused) / `enemy` (opponent left,
            win tainted). Drives W/L/D tally exclusion via the
            FilterRail leaver-handling setting.
          - **note** — free-text per-match commentary.
          - **replay_code** — Overwatch's six-character replay ID so
            the match can be re-watched in the in-game viewer.
          - **members** — BattleTags of group members; verbatim, no
            canonicalisation.

        Every field is optional, but at least one must carry content:
        this is an **upsert only**. An all-empty body is rejected with
        409 (the body parses but there's nothing to persist) — clearing
        an annotation is an explicit `DELETE` on this same path, so the
        verb states the intent.

        Returns 204 on success, 400 on a malformed body or a `leaver`
        outside the enum, 409 on an all-empty body, 500 on store error.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                # `note` and `replay_code` treat `null` and `""` as
                # the same "no value" — they're optional and the
                # server decodes either form as empty. `leaver` keeps
                # the strict enum because Spectral can't validate a
                # null member mixed into an enum (the parser bails);
                # the handler still tolerates `null` (Go decodes to
                # "" which is the explicit "no leaver tag" enum
                # value), so leaver: null lands as 204 — same outcome
                # as leaver: "".
                leaver:
                  $ref: "#/components/schemas/LeaverEnum"
                  description: |
                    One of the three leaver scenarios, or `""` to leave
                    untagged. An all-empty body (every field blank) is
                    rejected with 409 — use `DELETE` to clear the row.
                note:
                  type: string
                  description: Free-text per-match commentary. `null` = empty.
                replay_code:
                  type: string
                  description: Overwatch six-character replay ID. No format validation server-side. `null` = empty.
                members:
                  type: array
                  description: |
                    BattleTags of group members for this match. The server
                    trims whitespace, drops empty strings, and deduplicates;
                    the composite-PK on the child table catches any
                    duplicates that slip through. `null` is treated as an
                    empty list (clear members) — equivalent to `[]`.
                  items:
                    type: string
                    example: "Apollo#11234"
                tags:
                  type: array
                  description: |
                    Free-form labels applied to the match. `stack`,
                    `stream`, and `placement` are the conventional
                    three (the UI surfaces them as quick-add toggles)
                    but any string is accepted. Server lowercases +
                    trims + deduplicates before persistence so
                    `Stack` and `stack` collapse to one entry. Drives
                    the FilterRail "Tags" multi-select. `null` is
                    treated as an empty list (clear tags) — equivalent
                    to `[]`.
                  items:
                    type: string
                    example: stack
      responses:
        "204":
          description: Annotation persisted.
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: DeleteMatchAnnotation
      tags: [Matches]
      summary: Clear a per-match user annotation
      description: |
        Removes the match's annotation row entirely, cascading any
        member and tag rows away. Idempotent — deleting a match that has
        no annotation still returns 204. This is the explicit clear verb;
        `PUT` is upsert-only and rejects an all-empty body.
      responses:
        "204":
          description: Annotation cleared (or already absent).
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/matches/{match_key}/data:
    parameters:
      - $ref: "#/components/parameters/MatchKey"
    put:
      operationId: UpdateMatchData
      tags: [Matches]
      summary: Replace a match's user-data override set
      description: |
        Upserts the user override layer for a match — the editable copy
        kept separate from the parsed OCR rows. The body is the FULL
        current override set; a per-field revert is the same request with
        that field omitted. Overridden fields surface in the record's
        `edited_fields` and flip `source` to `ocr_edited`. Idempotent.

        Reset to pure OCR with the matching `DELETE`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserMatchDataInput"
      responses:
        "204":
          description: Override set persisted.
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: A map or hero override isn't in the Overwatch roster.
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: ResetMatchData
      tags: [Matches]
      summary: Reset a match to pure OCR (clear overrides)
      description: |
        Clears the user override set for a match, reverting an edited OCR
        match to its parsed values. (Deleting a hand-entered match is the
        per-match `DELETE /matches/{match_key}` hard-delete instead.)
        Idempotent — clearing an unedited match succeeds.
      responses:
        "204":
          description: Overrides cleared.
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/matches/{match_key}/review:
    parameters:
      - $ref: "#/components/parameters/MatchKey"
    put:
      operationId: SetMatchReview
      tags: [Matches]
      summary: Tag a match as reviewed
      description: |
        Records who reviewed this match's VOD. Two states:

          - `self`  — the user reviewed the VOD themselves
          - `coach` — a coach reviewed the VOD with the user

        Absence of a `match_reviews` row is the third logical state
        ("not reviewed"); use `DELETE` on the same path to revert
        there. Idempotent — repeated identical calls succeed.

        Returns 204 on success, 400 on validation failure
        (`reviewed_by` outside the enum), 500 on store error.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reviewed_by]
              properties:
                reviewed_by:
                  type: string
                  enum: [self, coach]
                  description: |
                    Who reviewed this match's VOD. To revert to "not
                    reviewed", call DELETE on this path instead of
                    PUTing an empty string.
      responses:
        "204":
          description: Review status updated.
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: ClearMatchReview
      tags: [Matches]
      summary: Clear the review-status tag
      description: |
        Removes the `match_reviews` row for this match, reverting it
        to the implicit "not reviewed" state. Idempotent — clearing
        an unreviewed match returns 204.
      responses:
        "204":
          description: Review tag cleared (or already absent).
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/matches/{match_key}/queue:
    parameters:
      - $ref: "#/components/parameters/MatchKey"
    put:
      operationId: SetMatchQueue
      tags: [Matches]
      summary: Override a match's queue type
      description: |
        Overrides the queue format auto-detected from the teams
        players-per-team count. Two states:

          - `role` — 5v5 role queue (locked 1-2-2 composition)
          - `open` — 6v6 open queue (any composition)

        A stored override always wins over detection. `DELETE` on the
        same path removes it, reverting the match to the detected value
        (or "queue not set" when nothing was detected). Idempotent —
        repeated identical calls succeed.

        Drives the radiogroup at the top of the match-detail panel
        and the Queue chip in "Narrow this set."

        Returns 204 on success, 400 on validation failure
        (`queue_type` outside the enum), 500 on store error.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [queue_type]
              properties:
                queue_type:
                  $ref: "#/components/schemas/QueueTypeEnum"
      responses:
        "204":
          description: Queue type updated.
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: ClearMatchQueue
      tags: [Matches]
      summary: Clear the queue-type tag
      description: |
        Removes the `match_queue` row for this match, reverting it
        to the implicit "queue not set" state. Idempotent — clearing
        an unset match returns 204.
      responses:
        "204":
          description: Queue tag cleared (or already absent).
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  parameters:
    MatchKey:
      in: path
      name: match_key
      required: true
      schema:
        type: string
        minLength: 1
      description: |
        Match identity — same `match_key` value exposed in
        `MatchRecord`. URL-safe: the canonical form replaces every
        legacy colon separator with a dash, so no percent-encoding
        is required for paste-in-URL use
        (e.g. `match-2026-05-10T22-21-11`). For `unmatched-<filename>`
        and `ambiguous-<filename>` variants the embedded filename
        still needs the usual encoding for spaces / unicode.
      example: "match-2026-05-10T22-21-11"
  schemas:
...
    LeaverEnum:
      type: string
      enum: ["", "self", "team", "enemy"]
      description: |
        Per-match leaver tag — who left the match.
        - `self`: the user left (their data is incomplete).
        - `team`: an ally left.
        - `enemy`: an opponent left.
        - `""` (empty string): no leaver tag set; functionally
          identical to omitting the field. Sending the empty string
          to the single-set endpoint is equivalent to a DELETE on
          the field.
    QueueTypeEnum:
      type: string
      enum: ["role", "open"]
      description: |
        Queue format the match was played in. The single-set
        `PUT /matches/{match_key}/queue` endpoint only accepts
        non-empty values; revert to "queue not set" via the matching
        DELETE.
...
  responses:
    BadRequest:
      description: Malformed request body or query parameters.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    NotFound:
      description: The requested resource was not found.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...
    InternalError:
      description: Unhandled server-side error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
...